Moved subcommands out of their package and renamed them accordingly
