+++ src/main/java/net/javadiscord/javabot/systems/help/commands/HelpCommand.java
@@ -2,9 +2,6 @@ package net.javadiscord.javabot.systems.help.commands;
 
 import com.dynxsty.dih4jda.interactions.commands.SlashCommand;
 import net.dv8tion.jda.api.interactions.commands.build.Commands;
-import net.javadiscord.javabot.systems.help.commands.subcommands.HelpAccountSubcommand;
-import net.javadiscord.javabot.systems.help.commands.subcommands.HelpGuidelinesCommand;
-import net.javadiscord.javabot.systems.help.commands.subcommands.HelpPingCommand;
 
 /**
  * Represents the `/help` command. This holds commands related to the help system.
@@ -17,6 +14,6 @@ public class HelpCommand extends SlashCommand {
 		setSlashCommandData(Commands.slash("help", "Commands related to the help system.")
 				.setGuildOnly(true)
 		);
-		addSubcommands(new HelpAccountSubcommand(), new HelpPingCommand(), new HelpGuidelinesCommand());
+		addSubcommands(new HelpAccountSubcommand(), new HelpPingSubcommand(), new HelpGuidelinesSubcommand());
 	}
 }
